requirements.txt
pipecat-ai
aiohttp
loguru
pandas
openpyxl
websockets
pydub
